Moving _metrics.py to results folder
